+++ android/src/newarch/java/io/sentry/react/RNSentryModule.java
@@ -2,10 +2,12 @@ package io.sentry.react;
 
 import androidx.annotation.NonNull;
 
+import com.facebook.react.bridge.JavaScriptExecutorFactory;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.Promise;
+import com.facebook.react.bridge.WritableMap;
 
 public class RNSentryModule extends NativeRNSentrySpec {
 
@@ -120,5 +122,15 @@ public class RNSentryModule extends NativeRNSentrySpec {
     @Override
     public void fetchNativeSdkInfo(Promise promise) {
         this.impl.fetchNativeSdkInfo(promise);
+    }
+
+    @Override
+    public WritableMap startProfiling() {
+        return this.impl.startProfiling();
+    }
+
+    @Override
+    public WritableMap stopProfiling() {
+        return this.impl.stopProfiling();
     }
 }