+++ test/mockDsn.ts
@@ -1,0 +1,1 @@
+export const MOCK_DSN = 'https://6890c2f6677340daa4804f8194804ea2@o19635.ingest.sentry.io/148053';